app/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
aiohttp==3.10.8
pandas==2.2.3
djangorestframework==3.15.2
djangorestframework-filters==1.0.0.dev0
six==1.16.0




